Uci (#19)
* Move cli chess board under `play` subcommand

Prepping for a better engine architecture. The simbelmyne crate will
house both the cli player, the perft tooling, and the engine (accessible
through a UCI interface on stdin/stdout)

* Basic UCI interface

Have a listener waiting for input on the main thread, and a secondary
worker thread that will do the searches. Still need to implement a bunch
of the options, but we've got something going, at least.