Decouple rendering interval from messages (#69)
At times, the engines churn through thousands of plys per second, (when
there's only a handful of nodes in the search tree). This means we're
getting spammed with `info` messages, that we're reading one at a time,
and doing a render pass every time we process a message.

Instead, we should set up a fixed FPS interval and include it in the
`select!` race. That way, we only render at fixed intervals, and process
all the messages we're receiving in between.