Position moves (#81)
* Have Position wrap a board and a vec of moves

* I take it back

* Add parsing of additional moves

Opted not to store the moves in the `UciClientMessage`.

I'm not entirely sure if that's the right call or not. It feels a bit
iffy for the UCI helpers to be playing moves on a board, etc...
Definitely means it's less useful to other people (let's be real, what
other people?) who might want access to the board position and moves
separately.

Maybe it was the wrong call after all?