Position moves (#83)
* Have Position wrap a board and a vec of moves

* I take it back

* Add parsing of additional moves

Opted not to store the moves in the `UciClientMessage`.

I'm not entirely sure if that's the right call or not. It feels a bit
iffy for the UCI helpers to be playing moves on a board, etc...
Definitely means it's less useful to other people (let's be real, what
other people?) who might want access to the board position and moves
separately.

Maybe it was the wrong call after all?

* Add helpers for matching an algebraic move to a legal move

Moves parsed from algebraic don't hold any metadata (like capture,
en-passant, etc), so we can't use simple equality.

But then, also can't just do a dumb src/tgt check, because that would
ignore promotion information (which we _do_ get from algebraic notation)

* Actually play the _full_ legal moves

Rather than the pseudo moves we get from algebraic.

Wonder if it makes sense to create a `RawMove` type that doesn't hold
any of the `Move` metadata, but we can implement equality between a
`Move` and a `RawMove`.