Add BareMove (#90)
Often, we parse moves we get over UCI from algebraic strings. These are
missing a lot of metadata that's encoded in our Move type.

Adding a BareMove type that acts as an intermediary. It can be
compared for equality with a Move, and it can be played to a board
position just fine (panics if the move is illegal, though!), and
parsed/printed from/to algebraic like regular moves.