Use `search.should_continue()` rather than `search.tc.stopped()`
`search.stopped()` actually just checks the global stop flag.

Having said that, `should_continue()` also checks the stopped flag, so
we'll want to start paying extra attention there, lest locking overhead
of the AtomicBool becomes a thing.

Most people seem to only really check every 2048 nodes or so. That is,
have the TC keep track of the nodecount when we last checked in, and
only do the actual checks if we exceed that window.