Don't copy conthist tables before passing them into `MovePicker::next`
Not _really_ a bugfix, but very much _not_ what I was intending.
Right now, we just copy the entire table and pass a reference into the
move picker. What we _really_ want is just a reference to the one, true,
conthist table.

The main difference here, apart from performance (?), is that when the
move picker actually uses the conthist table (inside `score_quiets()`),
we'll have the most up to date history (the values will be different,
because we'll have searched all of the tacticals and their entire
subtrees)

bench 7515987