Add `Position::approx_hash_after` helper (#266)
* Add `rough_hash_after` helper

Add helper that gives an approximate guess of the zobrist hash after
applying a certain move. It's used so we have a cheap way of getting the
zobrist hash ahead of time so we can prefetch the TT entry earlier.

Even if this returns the wrong hash 5% of the time, it just means that
we lose a marginal amount of pre-fetching, but are able to pre-fetch
_much_ earlier.

* Prefetch earlier in negamax and qsearch

Without changing bench nps! 🥳

bench 7202707