Update WDL model scaling
slightly less broken now?

Turns out that, if you don't pass a `NormalizeToPawnValue`, or a set of
`NormalizeData` from a previous run, it will just pick the stockfish
default values, which in no way apply to me. (The numbers they returned
were operating under the assumption that a pawn was normalized to 270 or
something, where it's closer to 100 for us)

Still not sure what the correct move there is, honestly. I just yolo'd a
value of 76.

bench 6183120